# docs/swagger.yml
swagger: "2.0"
info:
  title: "pms api"
  version: "1.0"
basePath: "/api"
paths:
...
  /project/{project_id}/users:
    get:
      summary: 查询项目成员
      operationId: controller.activity.project_user
      parameters:
        - name: project_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: ok
  /demand:
    get:
      summary: 查询需求
      operationId: controller.activity.demand_search
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - name: title
          in: query
          required: true
          type: string
        - name: projectId
          in: query
          required: true
          type: integer
      responses:
        200:
          description: ok
  /activity:
    post:
      summary: 创建活动
      operationId: controller.activity.activity_add
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - name: activity
          in: body
          description: user to add to the system
          required: true
          schema:
            $ref: '#/definitions/Activity'
      responses:
        200:
          description: 活动创建成功
...
  /project/{project_id}/user:
    get:
      summary: 查询项目成员
      operationId: controller.project.project_user
      parameters:
        - name: project_id
          in: path
          required: true
          type: number
      responses:
        200:
          description: ok
    post:
      summary: 更新项目成员
      operationId: controller.project.project_user_update
      parameters:
        - name: project_id
          in: path
          required: true
          type: number
      responses:
        200:
          description: ok
# @author: Xing Dong
...
  /demand/{demand_title}:
    get:
      summary: api
      operationId: controller.task.demand_search
      parameters:
        - name: demand_title
          in: path
          required: true
          type: string
      responses:
        200:
          description: ok
...
definitions:
  Activity:
    type: object
    properties:
      title:
        type: string
      projectId:
        type: integer
      demand:
        type: array
        items:
          type: integer
    required:
    - title
    - projectId
    - demand